Update to django-allauth 0.55.2, with changes
With this update, a provider that has a database-backed SocialApp (for
example, for OpenID client credentials) needs to be initialized with
that SocialApp. Switch to using the socialaccount app adapter to load
the FirefoxAccountProvider, and setup the SocialApp in tests.

With this update, non-null provider_id and settings fields are added to
the socialaccount_socialapp table. This is OK as long as a new social
app is not added during a database migration. For Relay testing, add a
migration with default values.